+++ src/transport/xqc_frame_parser.h
@@ -36,7 +36,7 @@ ssize_t xqc_gen_stream_frame(xqc_packet_out_t *packet_out,
 xqc_int_t xqc_parse_stream_frame(xqc_packet_in_t *packet_in, xqc_connection_t *conn,
     xqc_stream_frame_t *frame, xqc_stream_id_t *stream_id);
 
-ssize_t xqc_gen_crypto_frame(xqc_packet_out_t *packet_out, size_t offset,
+ssize_t xqc_gen_crypto_frame(xqc_packet_out_t *packet_out, uint64_t offset,
     const unsigned char *payload, uint64_t payload_size, size_t *written_size);
 
 xqc_int_t xqc_parse_crypto_frame(xqc_packet_in_t *packet_in, xqc_connection_t *conn, xqc_stream_frame_t * frame);
